Moved user events from events viewsets to profiles viewsets. Also corrected user events tests. Needs tests for all remaining endpoints
