+++ apps/conftest.py
@@ -1,7 +1,47 @@
-import pytest
+import os
+import django
 
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
+django.setup()
+
+import pytest
+from rest_framework.test import APIClient
+
+from apps.profiles.models import User
 from apps.profiles.models.followers import Follower
 
+
+@pytest.fixture
+def api_client() -> APIClient:
+    return APIClient()
+
+
+@pytest.fixture
+def user() -> User:
+    return User.objects.create(email="user@example.com", password="test")
+
+
+@pytest.fixture
+def user_2() -> User:
+    return User.objects.create(email="user2@example.com", password="test2")
+
+
+@pytest.fixture
+def user_private() -> User:
+    return User.objects.create(
+        email="user_private@example.com",
+        password="test",
+        is_private=True,
+    )
+
+
+@pytest.fixture
+def user_2_private() -> User:
+    return User.objects.create(
+        email="user2_private@example.com",
+        password="test2",
+        is_private=True,
+    )
 
 @pytest.fixture
 def follower_user_accepted(user, user_2) -> Follower:
