+++ apps/events/tests/events/constants.py
@@ -1,2 +1,2 @@
-CREATED_BY_URL = "events:User events-event_list_created_by_user"
-IS_PARTICIPANT_URL = "events:User events-event_list_user_is_participant"
+CREATED_BY_URL = "profiles:User events-event_list_created_by_user"
+IS_PARTICIPANT_URL = "profiles:User events-event_list_user_is_participant"
