+++ apps/events/urls.py
@@ -8,7 +8,6 @@ from .viewsets import (
     MarkerViewSet,
     CategoryViewSet,
     ReviewViewSet,
-    UserEventViewSet,
 )
 
 
@@ -17,7 +16,6 @@ app_name = "events"
 router = routers.SimpleRouter()
 router.register("events", EventViewSet, basename="Events")
 router.register("events/(?P<event_id>[^/.]+)/ratings", RatingViewSet, basename="Events rating")
-router.register("events", UserEventViewSet, basename="User events")
 router.register("categories", CategoryViewSet, basename="Categories")
 router.register("tags", TagViewSet, basename="Tags")
 router.register("markers", MarkerViewSet, basename="Markers")
