+++ apps/events/viewsets.py
@@ -196,71 +196,6 @@ class EventViewSet(viewsets.ModelViewSet):
         return Response(status=status.HTTP_200_OK)
 
 
-class UserEventViewSet(viewsets.ModelViewSet):
-    model = Event
-    serializer_class = EventListSerializer
-    permission_classes = [
-        IsAuthenticatedOrReadOnly,
-        IsAuthenticated,
-        FollowerPermissions,
-    ]
-    http_method_names = ["get"]
-    lookup_url_kwarg = "user_id"
-
-    def get_queryset(self):
-        self.queryset = (
-            self.model.objects
-            .annotate(participants_number=Count("participants"))
-            .order_by("-start_date")
-        )
-        return self.queryset.all()
-
-    @swagger_auto_schema(
-        request_body=no_body
-    )
-    @action(
-        methods=["get"],
-        detail=True,
-        permission_classes=[IsAuthenticatedOrReadOnly],
-        url_path="created_by",
-        url_name="event_list_created_by_user",
-    )
-    def list_created_by_user(self, request, user_id):
-        user = get_user_object(user_id)
-        queryset = self.get_queryset().filter(created_by=user)
-
-        if user != request.user:
-            queryset = queryset.filter(is_visible=True)
-
-        serializer = self.get_serializer(queryset, many=True)
-        return Response(serializer.data)
-
-    @action(
-        methods=["get"],
-        detail=True,
-        permission_classes=[IsAuthenticated, FollowerPermissions],
-        url_path="is_participant",
-        url_name="event_list_user_is_participant",
-    )
-    def list_user_is_participant(self, request, user_id):
-        user = get_user_object(user_id)
-        queryset = self.get_queryset().filter(participants__id=user_id)
-
-        if user != request.user:
-            queryset = queryset.filter(is_visible=True)
-
-        serializer = self.get_serializer(queryset, many=True)
-        return Response(serializer.data)
-
-    @swagger_auto_schema(auto_schema=None)
-    def retrieve(self, request):
-        pass
-
-    @swagger_auto_schema(auto_schema=None)
-    def list(self, request):
-        pass
-
-
 class RatingViewSet(viewsets.ModelViewSet):
     """
     A simple ViewSet for viewing and editing event ratings.
