+++ apps/profiles/urls.py
@@ -5,7 +5,7 @@ from apps.profiles.viewsets import (
     ProfileViewSet,
     UserRatingViewSet,
     MyProfileViewSet,
-    FollowerViewSet,
+    FollowerViewSet, ProfileEventViewSet,
 )
 
 app_name = "profiles"
@@ -14,6 +14,8 @@ router = routers.SimpleRouter()
 router.register('users/(?P<user_id>[^/.]+)/user_ratings', UserRatingViewSet, basename="UserRating")
 router.register("users", ProfileViewSet, basename="Profiles")
 router.register("users", FollowerViewSet, basename="Followers")
+router.register("users", ProfileEventViewSet, basename="User events")
+
 
 urlpatterns = [
     path("api/v1/", include(router.urls)),
