+++ apps/profiles/viewsets.py
@@ -1,5 +1,6 @@
 import logging
 
+from django.db.models import Count
 from drf_yasg.utils import swagger_auto_schema, no_body
 from rest_framework import viewsets, generics, status
 from rest_framework.decorators import action
@@ -7,6 +8,8 @@ from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnl
 from rest_framework.response import Response
 
 from apps.core.utils import delete_image_if_exists, validate_city
+from apps.events.models import Event
+from apps.events.serializers import EventListSerializer
 from apps.profiles.models import UserRating, User
 from apps.profiles.models.followers import Follower
 from apps.profiles.permissions import (
@@ -253,3 +256,68 @@ class FollowerViewSet(viewsets.ModelViewSet):
         queryset = self.queryset.filter(follower=user, status=Follower.Status.ACCEPTED)
         serializer = self.get_serializer(queryset, many=True)
         return Response(status=status.HTTP_200_OK, data=serializer.data)
+
+
+class ProfileEventViewSet(viewsets.ModelViewSet):
+    model = Event
+    serializer_class = EventListSerializer
+    permission_classes = [
+        IsAuthenticatedOrReadOnly,
+        IsAuthenticated,
+        FollowerPermissions,
+    ]
+    http_method_names = ["get"]
+    lookup_url_kwarg = "user_id"
+
+    def get_queryset(self):
+        self.queryset = (
+            self.model.objects
+            .annotate(participants_number=Count("participants"))
+            .order_by("-start_date")
+        )
+        return self.queryset.all()
+
+    @swagger_auto_schema(
+        request_body=no_body
+    )
+    @action(
+        methods=["get"],
+        detail=True,
+        permission_classes=[IsAuthenticatedOrReadOnly],
+        url_path="events/created_by",
+        url_name="event_list_created_by_user",
+    )
+    def list_created_by_user(self, request, user_id):
+        user = get_user_object(user_id)
+        queryset = self.get_queryset().filter(created_by=user)
+
+        if user != request.user:
+            queryset = queryset.filter(is_visible=True)
+
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
+    @action(
+        methods=["get"],
+        detail=True,
+        permission_classes=[IsAuthenticated, FollowerPermissions],
+        url_path="events/is_participant",
+        url_name="event_list_user_is_participant",
+    )
+    def list_user_is_participant(self, request, user_id):
+        user = get_user_object(user_id)
+        queryset = self.get_queryset().filter(participants__id=user_id)
+
+        if user != request.user:
+            queryset = queryset.filter(is_visible=True)
+
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
+    @swagger_auto_schema(auto_schema=None)
+    def retrieve(self, request):
+        pass
+
+    @swagger_auto_schema(auto_schema=None)
+    def list(self, request):
+        pass
